--- a/src/ImageTracking.jl
+++ b/src/ImageTracking.jl
@@ -54,6 +54,4 @@ export
     ConvolutionImplementation,
     MatrixImplementation
 
-
-
 end