--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -7,4 +7,3 @@ include("lucas_kanade.jl")
 include("haar.jl")
 include("visualization.jl")
 include("error_evaluation.jl")
-