Hi! I was trying to use LK optical flow and noticed that it didn't give the correct results.
Either the flow was too big or its direction was incorrect. Additionally it was dropping a lot of points, marking them as bad.

#### Bugs

Upon investigation I've noticed two major bugs:

1. Incorrect pyramid level in the `determine_pyramid_coordinates` function.
    The indexing starts at `1`, but when computing pyramid point, it should start at `0`, so the first level does not change the coordinates, since it is the original level.
2. In the beginning of the optical flow there is the conversion of points and initial displacements from `(row, col)` to `(x, y)` format.
    However, all the subsequent functions operate in the `(row, col)` format.
    Simple way to check this is to visualize the gradient windows for the points.

#### Fixes:

Do tell me if you don't agree with some of the changes.

- Use the correct index for pyramid level.
- Drop the conversion from `(row, col)` to `(x, y)` and operate only in `(row, col)` format.
- Set default value for the `eigenvalue_threshold` parameter to `1e-4` as in [OpenCV](https://docs.opencv.org/4.5.2/dc/d6b/group__video__track.html#ga473e4b886d0bcc6b65831eb88ed93323).
- Remove `in_image` function, since `lies_in` is enough.
- Convert `while` loops to explicit expressions in `get_grid` function.
- Simplify displacement creation in tests.
- Use default parameters in tests for Lucas-Kanade algorithm.
- Replace tabs in certain places with spaces for consistency purposes.
- Make rows no more than 80 characters long & minor refactoring (this allows for split-screen editing 🙃).

#### TODOs:

- [ ] Add epsilon termination criteria as in [OpenCV LK](https://docs.opencv.org/4.5.2/d9/d5d/classcv_1_1TermCriteria.html#a56fecdc291ccaba8aad27d67ccf72c57a857609e73e7028e638d2ea649f3b45d5).
- [ ] Add ability to pass precomputed pyramids and image gradients.
      This can be useful, when computing flow for the image with more than one other image.
      Since this way we would need to compute them only once.
      I think about either adding them to keyword parameters, or creating `LKPyramid` struct that will contain both pyramids and gradients. Creating struct will decrease number of parameters in the function, which would make things easier.
- [ ] Add tests that fail the current LK version, but succeed on this PR.
- [ ] Bring back deleted comments (this was done for simplification).

#### Visual comparison:

<details>
<summary>Images</summary>

Starting with the simple case, where the figure moves only horizontally.

|Current|This PR|
|:-:|:-:|
|![lk-old-1](https://user-images.githubusercontent.com/17990405/124880432-f4eb7880-dfd6-11eb-8a7f-5fe88122371a.png)|![lk-new-1](https://user-images.githubusercontent.com/17990405/124880455-f9b02c80-dfd6-11eb-8d02-8015fb5e6afb.png)|

Car moves forward. Two adjacent frames in a video.

|Current|This PR|
|:-:|:-:|
|![lk-old-3](https://user-images.githubusercontent.com/17990405/124880624-2b28f800-dfd7-11eb-9c4b-8dafa97bed83.png)|![lk-new-3](https://user-images.githubusercontent.com/17990405/124880601-25331700-dfd7-11eb-8f91-b091e0503959.png)|

Basketball example. Motion is primarily in the hands and the body.

|Current|This PR|
|:-:|:-:|
|![lk-old-4](https://user-images.githubusercontent.com/17990405/124880847-61667780-dfd7-11eb-9e55-37676d1c9788.png)|![lk-new-4](https://user-images.githubusercontent.com/17990405/124880864-64616800-dfd7-11eb-8106-6af79dba5d58.png)|

Cup. Horizontal movement.

|Current|This PR|
|:-:|:-:|
|![lk-old-2](https://user-images.githubusercontent.com/17990405/124881324-dfc31980-dfd7-11eb-83bb-2f4bc863afea.png)|![lk-new-2](https://user-images.githubusercontent.com/17990405/124881340-e487cd80-dfd7-11eb-93b0-3d505ec15032.png)|

</details>