@zygmuntszpak 2021-07-08
Thanks for working on this. Looks very promising! I'll need a couple of days to reacquaint myself with the code and algorithm. There is definitely a problem with the current implementation because I can see in the old code that we have ``Ix, Iy = imgradients``  instead of ``Iy, Ix = imgradients`` which is one of the several things you have fixed!