@zygmuntszpak 2021-07-10
> Hi! I was trying to use LK optical flow and noticed that it didn't give the correct results.
> Either the flow was too big or its direction was incorrect. Additionally it was dropping a lot of points, marking them as bad.
> #### Bugs:
> 
> Upon investigation I've noticed two major bugs:
> 
>     1. Incorrect pyramid level in the `determine_pyramid_coordinates` function.
>        The indexing starts at `1`, but when computing pyramid point, it should start at `0`, so the first level does not change the coordinates, since it is the original level.
> 
>     2. In the beginning of the optical flow there is the conversion of points and initial displacements from `(row, col)` to `(x, y)` format.
>        However, all subsequent functions operate in the `(row, col)` format.
>        Simple way to check this is to visualize the gradient windows for the points.

Thanks for addressing these colossal bugs. 

> 
> #### Fixes:
> 
> Do tell me if you don't agree with some of the changes.
> 
>     * Use the correct index for pyramid level.
> 
>     * Drop the conversion from `(row, col)` to `(x, y)` and operate only in `(row, col)` format.
> 
>     * Set default value for the `eigenvalue_threshold` parameter to `1e-4` as in [OpenCV](https://docs.opencv.org/4.5.2/dc/d6b/group__video__track.html#ga473e4b886d0bcc6b65831eb88ed93323).
> 
>     * Remove `in_image` function, since `lies_in` is enough.
> 
>     * Convert `while` loops to explicit expressions in `get_grid` function.
> 
>     * Simplify displacement creation in tests.
> 
>     * Use default parameters in tests for Lucas-Kanade algorithm.
> 
>     * Replace tabs in certain places with spaces for consistency purposes.
> 
>     * Make rows no more than 80 characters long & minor refactoring (this allows for split-screen editing upside_down_face).

I think overall these are some excellent changes and I am grateful for the time you put into making them. 

If we are going to be changing the code style then I suggest we follow the [BlueStyle](https://github.com/invenia/BlueStyle) which I believe already has a lot of traction in the Julia community. One of the key implications is that the 80 character limit would become a 92 character limit. 



> 
> #### TODOs:
> 
>     * [ ]  Add epsilon termination criteria as in [OpenCV LK](https://docs.opencv.org/4.5.2/d9/d5d/classcv_1_1TermCriteria.html#a56fecdc291ccaba8aad27d67ccf72c57a857609e73e7028e638d2ea649f3b45d5).
> 
>     * [ ]  Add ability to pass precomputed pyramids and image gradients.
>       This can be useful, when computing flow for the image with more than one other image.
>       Since this way we would need to compute them only once.
>       I think about either adding them to keyword parameters, or creating `LKPyramid` struct that will contain both pyramids and gradients. Creating struct will decrease number of parameters in the function, which would make things easier.
> 
>     * [ ]  Add tests that fail the current LK version, but succeed on this PR.
> 
>     * [ ]  Bring back deleted comments (this was done for simplification).
> 
> 

Those are great suggestions. However, I propose you keep the "ability to pass precomputed pyramids and image gradients." for a separate pull request. 

One important issue that needs to be addressed in this PR is an apparent substantial increase in runtime. Using [BenchmarkTools](https://github.com/JuliaCI/BenchmarkTools.jl) I found that the existing code takes approximately 1 second to complete on a pair of test images on my machine, whereas this pull-request takes 5 seconds for the same pair. I believe the issue is that your proposed version allocates memory within performance-critical loops. I think this can be traced down to you removing the use of the custom pinv2x2 function in favour of the built-in pinv function. Revisting this code, I think that we may be better-off using the svd2x2 function directly. This will allow us to construct the pseudo-inverse, but also give us the singular values which we can use to define min_eigenvalue (if I recall correctly, for real symmetric matrices with non-negative eigenvalues, the eigenvalues and singular values are the same). 

Thanks again for reviving this implementation.