@pxl-th 2021-07-11
Hi! Thanks for the review. I've made several improvements and changes based on your comments.

One thing to note is that at the moment of your review I've already implemented version of the algorithm that operates on the `LKPyramid` struct that holds pyramid for the image and its gradients.
However, I've not exposed the functionality to the user, so from API standpoint it is still the same.
Only underlying functions are utilizing `LKPyramid` struct.
I'd appreciate if we can add these changes in this pull request. But we can expose the necessary API in the separate PR.

I've also used `svd2x2` to compute minimum singular values, since they indeed coincide with the eigenvalues in this case.

#### Quick summary of changes from previous commit
- Changed max line width to 92 symbols.
- Remove `is_truncated_window`.
  Use `get_grid` to determine whether we need to recompute gradient or can use the previous one.
- Remove `flow` vector, since it was used only to store pyramid layer contributions and was redundant.
- Remove extrapolation and use only interpolation for the second pyramid's layers.
  Since we never operate on the out-of-bounds regions.
- Use `svd2x2` to compute minimum eigenvalues instead of calling `eigen` (singular values conicide with eigenvalues in this case).
- Add epsilon termination criteria. Add the necessary parameter to `LucasKanade` struct as well.
- Update LK tests to require higher accuracy.

#### Benchmarking results (on 1000 initial keypoints)

Current:
`1.555 s (866161 allocations: 260.38 MiB)`

This PR:
- With epsilon termination criteria (used by default): `458.365 ms (7073 allocations: 240.53 MiB)`
- Without epsilon termination criteria: `1.827 s (7091 allocations: 240.53 MiB)`
- With precomputed pyramids: `57.039 ms (10 allocations: 10.31 MiB)`

Note that without epsilon termination criteria the time is slightly bigger.
This can be because current version and this PR are tracking and invalidating different amount of points.