@pxl-th 2021-07-12
#### Benchmarking results with the latest changes

Parameters used:
- window size = 21
- levels = 3
- iterations = 30

Current:
`1.555 s (866161 allocations: 260.38 MiB)`

This PR:
- With epsilon termination criteria: `352.698 ms (6839 allocations: 199.00 MiB)`
- Using precomputed pyramids: `131.568 ms (10 allocations: 10.31 MiB)`